requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
streamlit
python-dotenv
sentence-transformers
